Move EnketoError exception to onadata.libs.exceptions.py
Moved this exception here for consistency..
