+++ onadata/libs/serializers/xform_serializer.py
@@ -23,7 +23,8 @@ from onadata.libs.utils.cache_tools import (
 from onadata.libs.utils.common_tags import GROUP_DELIMETER_TAG
 from onadata.libs.utils.decorators import check_obj
 from onadata.libs.utils.viewer_tools import (
-    EnketoError, enketo_url, get_enketo_preview_url, get_form_url)
+    enketo_url, get_enketo_preview_url, get_form_url)
+from onadata.libs.exceptions import EnketoError
 
 
 def _create_enketo_url(request, xform):
